+++ examples/react/src/main.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "@agile-team/mach-table/styles/mach-table.css";
+import "@agile-team/mach-table-react/styles.css";
 import App from "./App";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(