+++ examples/vue/src/main.ts
@@ -1,5 +1,5 @@
 import { createApp } from "vue";
-import "@agile-team/mach-table/styles/mach-table.css";
+import "@agile-team/mach-table-vue/styles.css";
 import App from "./App.vue";
 
 createApp(App).mount("#app");