+++ packages/react/src/index.ts
@@ -5,18 +5,4 @@ export { useMachGrid } from "./useMachGrid";
 export type { UseMachGridReturn } from "./useMachGrid";
 export { reactCellRenderer, reactDetailRenderer } from "./adapters";
 export type { ReactCellRendererProps, ReactDetailProps } from "./adapters";
-
-export type {
-  GridApi,
-  GridOptions,
-  ColDef,
-  RowNode,
-  CellRendererParams,
-  CellClickEvent,
-  SelectionChangedEvent,
-  CellValueChangedEvent,
-  SortChangedEvent,
-  FilterChangedEvent
-} from "@agile-team/mach-table";
-
-export { createGrid, EVENT_TYPES } from "@agile-team/mach-table";
+export * from "@agile-team/mach-table";